+++ src/assets/jss/material-kit-react/components/customLinearProgressStyle.jsx
@@ -1,7 +1,3 @@
-// ##############################
-// // // CustomLinearProgress component styles
-// #############################
-
 import {
   primaryColor,
   warningColor,