+++ src/assets/jss/material-kit-react/components/infoStyle.jsx
@@ -1,7 +1,3 @@
-// ##############################
-// // // Info component styles
-// #############################
-
 import {
   primaryColor,
   warningColor,