+++ src/components/Card/CardHeader.jsx
@@ -11,11 +11,11 @@ import withStyles from "@material-ui/core/styles/withStyles";
 import cardHeaderStyle from "assets/jss/material-kit-react/components/cardHeaderStyle.jsx";
 
 function CardHeader({ ...props }) {
-  const { classes, className, children, color, plainCard, ...rest } = props;
+  const { classes, className, children, color, plain, ...rest } = props;
   const cardHeaderClasses = classNames({
     [classes.cardHeader]: true,
     [classes[color + "CardHeader"]]: color,
-    [classes.cardHeaderPlain]: plainCard,
+    [classes.cardHeaderPlain]: plain,
     [className]: className !== undefined
   });
   return (
@@ -29,7 +29,7 @@ CardHeader.propTypes = {
   classes: PropTypes.object.isRequired,
   className: PropTypes.string,
   color: PropTypes.oneOf(["warning", "success", "danger", "info", "primary"]),
-  plainCard: PropTypes.bool
+  plain: PropTypes.bool
 };
 
 export default withStyles(cardHeaderStyle)(CardHeader);