+++ src/components/CustomTabs/CustomTabs.jsx
@@ -4,103 +4,77 @@ import classNames from "classnames";
 // nodejs library to set properties for components
 import PropTypes from "prop-types";
 
-// @material-ui/core components
+// material-ui components
 import withStyles from "@material-ui/core/styles/withStyles";
 import Tabs from "@material-ui/core/Tabs";
 import Tab from "@material-ui/core/Tab";
-
 // core components
 import Card from "components/Card/Card.jsx";
 import CardBody from "components/Card/CardBody.jsx";
 import CardHeader from "components/Card/CardHeader.jsx";
+
 import customTabsStyle from "assets/jss/material-kit-react/components/customTabsStyle.jsx";
 
 class CustomTabs extends React.Component {
   state = {
     value: 0
   };
+
   handleChange = (event, value) => {
     this.setState({ value });
   };
+
   render() {
     const {
       classes,
       headerColor,
-      title,
+      plainTabs,
       tabs,
-      rtlActive,
-      plainTabs
+      title,
+      rtlActive
     } = this.props;
     const cardTitle = classNames({
       [classes.cardTitle]: true,
       [classes.cardTitleRTL]: rtlActive
     });
-    const tabsContainer = classNames({
-      [classes.tabsContainer]: true,
-      [classes.tabsContainerRTL]: rtlActive
-    });
     return (
       <Card plain={plainTabs}>
-        <CardHeader color={headerColor} plainCard={plainTabs}>
+        <CardHeader color={headerColor} plain={plainTabs}>
           {title !== undefined ? (
-            <div className={cardTitle}>{"title"}</div>
+            <div className={cardTitle}>{title}</div>
           ) : null}
           <Tabs
-            classes={{
-              root: classes.customTabsRoot,
-              flexContainer: tabsContainer,
-              indicator: classes.displayNone
-            }}
             value={this.state.value}
             onChange={this.handleChange}
-            textColor="inherit"
+            classes={{
+              root: classes.tabsRoot,
+              indicator: classes.displayNone
+            }}
           >
             {tabs.map((prop, key) => {
               var icon = {};
-              if (prop.tabIcon !== undefined) {
+              if (prop.tabIcon) {
                 icon = {
-                  icon: <prop.tabIcon className={classes.tabIcon} />
+                  icon: <prop.tabIcon />
                 };
-              } else {
-                icon = {};
               }
               return (
                 <Tab
-                  key={key}
                   classes={{
-                    root: classes.customTabRoot,
-                    selected: classes.customTabSelected,
-                    labelContainer: classes.customTabLabelContainer,
-                    wrapper: classes.customTabWrapper,
-                    label: classes.customTabLabel
-                    // wrapper: tabWrapper,
-                    // labelIcon: labelIcon,
-                    // label: classes.label,
-                    // textColorInherit: labelIcon,
-                    // // textColorInheritSelected: classes.textColorInheritSelected + " " + labelIcon,
-                    // selected: classes.textColorInheritSelected + " " + labelIcon,
-                    // labelContainer: labelContainer
+                    root: classes.tabRootButton,
+                    labelContainer: classes.tabLabelContainer,
+                    label: classes.tabLabel,
+                    selected: classes.tabSelected,
+                    wrapper: classes.tabWrapper
                   }}
-                  // icon={<prop.tabIcon className={tabIcon} />}
-                  {...icon}
+                  key={key}
                   label={prop.tabName}
+                  {...icon}
                 />
               );
             })}
           </Tabs>
         </CardHeader>
-        {/* <CardHeader
-          classes={{
-            root: cardHeader,
-            title: cardTitle,
-            content: classes.cardHeaderContent,
-            action: classes.cardHeaderAction
-          }}
-          title={title}
-          action={
-
-          }
-        /> */}
         <CardBody>
           {tabs.map((prop, key) => {
             if (key === this.state.value) {
@@ -113,10 +87,6 @@ class CustomTabs extends React.Component {
     );
   }
 }
-
-CustomTabs.defaultProps = {
-  headerColor: "purple"
-};
 
 CustomTabs.propTypes = {
   classes: PropTypes.object.isRequired,