+++ src/index.js
@@ -5,7 +5,7 @@ import { Router, Route, Switch } from "react-router";
 
 import indexRoutes from "routes/index.jsx";
 
-import "assets/scss/material-kit-react.css";
+import "assets/scss/material-kit-react.css?v=1.1.0";
 
 var hist = createBrowserHistory();